Further Optimizations
Moved chain logic back into one file with two classes
Simplified chat to simply chain together calls of think and response
fixed bug in `bot.py` for the restart method
