+++ www/main.py
@@ -4,7 +4,7 @@ import streamlit as st
 import time
 from agent.cache import LRUCache
 
-from agent.chain import chat, ConversationCache, load_chains
+from agent.chain import ConversationCache, load_chains
 import asyncio
 
 
@@ -73,24 +73,23 @@ for message in st.session_state.messages:
 
 
 thought, response = '', ''
-async def chat_and_save(local_chain: ConversationCache, input: str) -> tuple[str, str]:
+async def chat_and_save(local_chain: ConversationCache, input: str) -> None:
         global thought, response
         bloom_chain =  BLOOM_CHAIN # if local_chain.conversation_type == "discuss" else WORKSHOP_RESPONSE_CHAIN
         # response_chain = local_chain.conversation_type == "discuss" ? DISCUSS_RESPONSE_CHAIN : WORKSHOP_RESPONSE_CHAIN
-
-        thought = await chat(
-            inp=input,
-            thought_chain=bloom_chain,
-            thought_memory=local_chain.thought_memory
-        )
-        response = await chat(
-            inp=input,
-            thought=thought,
-            response_chain=bloom_chain,
-            response_memory=local_chain.response_memory
-        )
-
-        return
+        thought, response = await bloom_chain.chat(local_chain, input)
+        # thought = await chat(
+        #     inp=input,
+        #     thought_chain=bloom_chain,
+        #     thought_memory=local_chain.thought_memory
+        # )
+        # response = await chat(
+        #     inp=input,
+        #     thought=thought,
+        #     response_chain=bloom_chain,
+        #     response_memory=local_chain.response_memory
+        # )
+        return None
         
 
 if prompt := st.chat_input("hello!"):
